Add patterned inline-euclid arguments and mini-notation decimal tempo factors (#1412)
Two grammar-parity refinements toward TidalCycles:

Patterned inline euclid (Tidal bd("<3 5>", 8)): the inline euclid sugar
bd(pulses, steps[, rot]) now accepts cycle-varying number patterns for any
argument (bd(<3 5>, 8), bd(3, 8, <0 2>)). Each control is sampled once per
cycle at the cycle start, and cycle k gates as
euclid(pulses_k, steps_k, rotation_k) over the token repeated steps_k times
(PatternRuntime::EuclidPattern / query_euclid_pattern, value.rs). Constancy
is decided structurally (cycle_invariant_constant), so all-constant calls
keep the historical exact construction bit-for-bit. Per-cycle values share
the constant path's validation rules: unit-cycle values are validated
eagerly at eval time, later cycles when their cycle is queried (matching the
patterned tempo-factor precedent). Controls are sampled over full cycle
spans, so chunked queries equal whole-span queries.

Mini-notation decimal tempo factors: tight a*1.5 / a/1.5 step modifiers now
accept decimal literals, converting through the same exact
decimal-to-rational path and bounds as fast/slow
(positive_rational_tempo_factor: reduced numerator and denominator each at
most 1024). bd*0.5 equals bd/2 (Tidal semantics), StepOp::Slow now carries
the raw f64 like StepOp::Fast, and tight decimal * inside graph { ... } /
voice { ... } bodies still rewrites to pedal-DSL multiplication (dry*0.2,
x*1.5 covered by regression tests).

Roadmap: ticked the mini-notation integer-literal caveat and added the
patterned inline-euclid entry.

+++ crates/orpheus-lang/src/ast.rs
@@ -184,14 +184,17 @@ pub enum BinaryOp {
 /// ```
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum StepOp {
-    /// `a*n`: repeats the element `n` times within its own slot.
+    /// `a*n`: speeds the element up by `n` within its own slot (`bd*2`
+    /// repeats it twice, `bd*1.5` plays it at 3/2 speed).
     ///
     /// The raw factor is kept as parsed; inside `graph { ... }` blocks the
     /// parser rewrites it back into pedal-DSL multiplication, elsewhere it
-    /// must be an integer within `1..=1024`.
+    /// must convert to a positive rational whose numerator and denominator
+    /// are each at most 1024 (the shared `fast`/`slow` factor rules).
     Fast(f64),
-    /// `a/n`: slows the element so it takes `n` cycles to complete.
-    Slow(i64),
+    /// `a/n`: slows the element so it takes `n` cycles to complete; decimal
+    /// factors convert to exact rationals under the same bounds as `a*n`.
+    Slow(f64),
     /// `a!n`: replicates the element as `n` separate sequence steps.
     Replicate(i64),
     /// `a?` / `a?p`: randomly removes the element's events with drop
